Retrieve failure information from the error log when available
